lattice_north_to_qcl/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
oauthlib==3.2.2
pydantic==1.10.7
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
redis
redis-py-cluster
pymongo
python-decouple